Move session controllers into admin module namespace
The session and password controllers were never meant to also
handle frontend authentication.

Moving the controllers into the correct scope makes it possible
to write frontend controllers w/o Alchemys admins controllers
interfering.
