+++ app/views/alchemy/admin/user_sessions/new.html.erb
@@ -5,7 +5,7 @@
     <%= f.input :password %>
     <div class="input link">
       <small>
-        <%= link_to Alchemy.t('Forgot your password?'), new_password_path %>
+        <%= link_to Alchemy.t('Forgot your password?'), admin_new_password_path %>
       </small>
     </div>
     <div class="submit">
