+++ app/views/alchemy/notifications/alchemy_user_created.de.text.erb
@@ -10,6 +10,6 @@ Aus Sicherheitsgründen stellen wir Ihr Passwort hier nicht dar.
 
 Wenn Sie Ihr Passwort vergessen haben oder dies Ihr erster Login ist, gehen Sie bitte auf:
 
-<%= alchemy.new_password_url %>
+<%= alchemy.admin_new_password_url %>
 
 Viel Spaß mit Alchemy!
