+++ app/views/alchemy/notifications/alchemy_user_created.en.text.erb
@@ -10,6 +10,6 @@ For security reasons we do not show your password here.
 
 If you forgot your password or this is your first login, please goto:
 
-<%= alchemy.new_password_url %>
+<%= alchemy.admin_new_password_url %>
 
 Have much fun with Alchemy!
