+++ app/views/alchemy/notifications/alchemy_user_created.es.text.erb
@@ -10,6 +10,6 @@ Por razones de seguridad no mostraremos aquí tu contraseña.
 
 Si has olvidado tu contraseña o es tu primer acceso, por favor, accede a:
 
-<%= alchemy.new_password_url %>
+<%= alchemy.admin_new_password_url %>
 
 ¡Disfruta de tu web con Alchemy!
