+++ app/views/alchemy/notifications/alchemy_user_created.ru.text.erb
@@ -10,6 +10,6 @@
 
 Если вы забыли ваш пароль или если это ваш первый визит, пожалуйста, перейдите по ссылке:
 
-<%= alchemy.new_password_url %>
+<%= alchemy.admin_new_password_url %>
 
 Наслаждайтесь использованием Alchemy!
