+++ app/views/alchemy/notifications/member_created.de.text.erb
@@ -8,6 +8,6 @@ Es wurde ein Zufallspasswort generiert. Aus Sicherheitsgründen wird dieses hier
 
 Klicken Sie bitte daher auf den folgenden Link (*), um sich ein neues Passwort zu vergeben:
 
-<%= alchemy.new_password_url(email: @user.email) %>
+<%= alchemy.admin_new_password_url(email: @user.email) %>
 
 *) Sollte dies nicht funktionieren, so kopieren Sie Sich bitte diesen Link und tragen Sie ihn in die Adresszeile Ihres Webbrowsers ein.
