+++ app/views/alchemy/notifications/member_created.en.text.erb
@@ -8,6 +8,6 @@ The password was randomly created. For security reasons it is not displayed here
 
 Please follow this link (*) to reset your password:
 
-<%= alchemy.new_password_url(email: @user.email) %>
+<%= alchemy.admin_new_password_url(email: @user.email) %>
 
 *) If this does not work, please copy the url and paste it into the address bar of your web browser.
