+++ app/views/alchemy/notifications/member_created.es.text.erb
@@ -8,6 +8,6 @@ La contraseña se ha generado aleatoriamente. Por razones de seguridad no se mos
 
 Por favor, accede este enlace (*) para reiniciar tu contraseña:
 
-<%= alchemy.new_password_url(email: @user.email) %>
+<%= alchemy.admin_new_password_url(email: @user.email) %>
 
 *) Si esto no funciona, por favor, copia el enlace y pegalo en el campo de direcciones de tu navegador web.
