+++ app/views/alchemy/notifications/member_created.ru.text.erb
@@ -8,6 +8,6 @@
 
 Пожалуйста, пройдите по ссылке (*), чтобы назначить собственный пароль:
 
-<%= alchemy.new_password_url(email: @user.email) %>
+<%= alchemy.admin_new_password_url(email: @user.email) %>
 
 *) Если ссылка не работает, пожалуйста, скопируйте ее в адресную строку вашего браузера.
