+++ app/views/alchemy/notifications/reset_password_instructions.de.text.erb
@@ -2,7 +2,7 @@ Hallo <%= @user.fullname %>.
 
 Sie haben angefordert Ihr Passwort zurückzusetzen. Dies kann durch anklicken des nachfolgenden Links bestätigt werden.
 
-<%= alchemy.edit_password_url(@user, reset_password_token: @token) %>
+<%= alchemy.admin_edit_password_url(@user, reset_password_token: @token) %>
 
 Wenn Sie diese Zurücksetzung nicht angefragt haben, dann können Sie diese E-Mail einfach ignorieren.
 Ihr Passwort wird erst dann zurückgesetzt, wenn Sie den Link anklicken.
