+++ app/views/alchemy/notifications/reset_password_instructions.en.text.erb
@@ -2,7 +2,7 @@ Hello <%= @user.name %>.
 
 You have requested to change your password. Please confirm this by clicking the link below.
 
-<%= alchemy.edit_password_url(@user, reset_password_token: @token) %>
+<%= alchemy.admin_edit_password_url(@user, reset_password_token: @token) %>
 
 If you didn't request this, please ignore this email.
 Your password won't change until you access the link above and create a new one.
