+++ app/views/alchemy/notifications/reset_password_instructions.es.text.erb
@@ -2,7 +2,7 @@ Hola <%= @user.name %>.
 
 Has solicitado modificar tu contraseña. Por favor, confírmalo pulsando en el siguiente enlace.
 
-<%= alchemy.edit_password_url(@user, reset_password_token: @token) %>
+<%= alchemy.admin_edit_password_url(@user, reset_password_token: @token) %>
 
 Si no has sido tu el que ha hecho la solicitud, ignora este correo.
 Tu contraseña no cambiará hasta que no accedas al enlace de arriba y generes una nueva.
