+++ app/views/alchemy/notifications/reset_password_instructions.ru.text.erb
@@ -2,7 +2,7 @@
 
 Вы сделали запрос на смену пароля. Пожалуйста подтвердите это, нажав на ссылку ниже.
 
-<%= alchemy.edit_password_url(@user, reset_password_token: @token) %>
+<%= alchemy.admin_edit_password_url(@user, reset_password_token: @token) %>
 
 Если вы не делали запрос, просто проигнорируйте это письмо.
 Ваш пароль не изменится до тех пор, пока вы не перейдете по ссылке и сами не измените его.
