+++ spec/features/password_reset_feature_spec.rb
@@ -4,13 +4,13 @@ describe "Password reset feature." do
   let(:user) { create(:alchemy_admin_user) }
 
   it "User can visit password reset form." do
-    visit new_password_path
+    visit admin_new_password_path
 
     expect(page).to have_content('Password reset')
   end
 
   it "User can request password reset." do
-    visit new_password_path
+    visit admin_new_password_path
 
     fill_in :user_email, with: user.email
     click_button 'Send reset instructions'
@@ -24,7 +24,7 @@ describe "Password reset feature." do
       .to receive(:reset_password_by_token)
       .and_return(user)
 
-    visit edit_password_path(id: user.id, reset_password_token: '1234')
+    visit admin_edit_password_path(id: user.id, reset_password_token: '1234')
 
     fill_in :user_password, with: 'secret123'
     fill_in :user_password_confirmation, with: 'secret123'
