+++ spec/mailers/notifications_spec.rb
@@ -27,7 +27,7 @@ module Alchemy
       end
 
       it "mail body includes password instructions" do
-        expect(mail.body.raw_source).to match /#{Regexp.escape(new_password_url(email: user.email, only_path: true))}/
+        expect(mail.body.raw_source).to match /#{Regexp.escape(admin_new_password_url(email: user.email, only_path: true))}/
       end
     end
 
@@ -45,7 +45,7 @@ module Alchemy
       end
 
       it "mail body includes password instructions" do
-        expect(mail.body.raw_source).to match /#{Regexp.escape(new_password_url(only_path: true))}/
+        expect(mail.body.raw_source).to match /#{Regexp.escape(admin_new_password_url(only_path: true))}/
       end
     end
 
@@ -75,7 +75,7 @@ module Alchemy
       end
 
       it "mail body includes reset instructions" do
-        expect(mail.body.raw_source).to match /#{Regexp.escape(edit_password_url(user, reset_password_token: token, only_path: true))}/
+        expect(mail.body.raw_source).to match /#{Regexp.escape(admin_edit_password_url(user, reset_password_token: token, only_path: true))}/
       end
     end
   end
